Clarify the differences between Go and Rust in README.md
Coming from Go, Rust is a bit weird. It is very explicit about certain things, but also uses a lot of inference for other things. It seems that the pattern is "be explicit at boundaries, infer locally". So, it'll infer as much as it can, but if it can help you avoid mistakes, it'll require you to be explicit.